Renamed ConsecutiveStorageCanvas to InterleavedStorageCanvas to better match the DICOM vocabulary
