+++ lib/canvas/GrayscaleCanvas.js
@@ -1,6 +1,6 @@
 "use strict";
 
-const ConsecutiveStorageCanvas = require("./ConsecutiveStorageCanvas");
+const ConsecutiveStorageCanvas = require("./InterleavedStorageCanvas");
 
 module.exports = class GrayscaleCanvas extends ConsecutiveStorageCanvas {
