+++ lib/canvas/InterleavedStorageCanvas.js
@@ -8,7 +8,7 @@ const MultiValueStorageCanvas = require("./MultiValueStorageCanvas");
  *
  * The pixel value must be an array with the exact number of elements of that pixel.
  */
-module.exports = class ConsecutiveStorageCanvas extends MultiValueStorageCanvas {
+module.exports = class InterleavedStorageCanvas extends MultiValueStorageCanvas {
 
     _offset(x, y, component) {
         return (y * this._width + x) * this._bytesPerPixel * this._valuesPerPixel + component * this._bytesPerPixel;
