+++ lib/canvas/MultiValueStorageCanvas.js
@@ -2,7 +2,7 @@
 
 const Canvas = require("./Canvas");
 
-module.exports = class ConsecutiveStorageCanvas extends Canvas {
+module.exports = class MultiValueStorageCanvas extends Canvas {
 
     /**
      * A canvas that stores values with a fixed number of components (e.g. greyscale, RGB).
