+++ lib/canvas/RGBCanvas.js
@@ -1,6 +1,6 @@
 "use strict";
 
-const ConsecutiveStorageCanvas = require("./ConsecutiveStorageCanvas");
+const ConsecutiveStorageCanvas = require("./InterleavedStorageCanvas");
 
 module.exports = class RGBCanvas extends ConsecutiveStorageCanvas {
